backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
sqlalchemy
langchain-text-splitters
faiss-cpu
openai
numpy
